TrustID-DataEngineering/requirements.txt
sqlmodel==0.0.6
psycopg2-binary==2.9.3
indy-sdk==1.16.0
cryptography==36.0.1
requests==2.26.0
python-dotenv==0.19.2
pytest==7.0.1
redis==4.1.0